% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recocam_various_helper_functions.R
\name{fix_attribute_names}
\alias{fix_attribute_names}
\title{Returns the Exifs attribute name from the index in the researcher tagged input document.
This function may be deprecated; consider checking?}
\usage{
fix_attribute_names(keyword_id, keywords_tags)
}
\arguments{
\item{keyword_id}{Attribute index in the researcher tagged input document}

\item{keywords_tags}{data.frame of the exifs tag index to attribute mapping present in the image.}
}
\value{
Returns the attribute keyword from the index.
}
\description{
Returns the Exifs attribute name from the index in the researcher tagged input document.
This function may be deprecated; consider checking?
}
