% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recocam_camtrapr_tables.R
\name{make_effortstable}
\alias{make_effortstable}
\title{Generates a dataframe that summarizes the operational days of each camera used in the trip.}
\usage{
make_effortstable(
  data_from_images,
  camera_stations_file_path,
  trip,
  ct_columns = c("trip", "treatment", "site_code", "site", "camera_no",
    "date_time_original", "species", "comments")
)
}
\arguments{
\item{data_from_images}{dataframe with the extracted exif data with key attributes, Trip, DateTimeOriginal, Camera.No, Species, Count.}

\item{camera_stations_file_path}{Path string to the  to camera stations csv file that has 5 attributes, SiteCode, lat, long,
Site, Camera_Trap, Notes, Easting, Northing.}

\item{trip}{Character string representing the trip. E.g. Jun.10, Jun.11, Sep.14, Apr.15. The value is the trip for which the camera efforts are to be calculated.}

\item{ct_columns}{A list of columns to target in the data.frame of extracted camera trap data.}
}
\value{
Returns a dataframe compatabile with CTtable of camtrapR package.
}
\description{
Generates a dataframe that summarizes the operational days of each camera used in the trip.
}
