% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recocam_camtrapr_tables.R
\name{make_recordtable}
\alias{make_recordtable}
\title{Generate a data frame that is compatible with the camtrapR recordTable data structure.}
\usage{
make_recordtable(
  data_from_images,
  delta_time_compared_to = "lastIndependentRecord",
  camera_independence = TRUE,
  remove_non_independent_records = TRUE,
  min_delta_time = 3
)
}
\arguments{
\item{data_from_images}{dataframe of image data created using the gather_images function}

\item{delta_time_compared_to}{Character string having values of lastIndependentRecord (default) or lastRecord. This is used to compare the time difference to previous records of same species in a station.}

\item{camera_independence}{logical. option to control if cameras are independent from each other. Default is set to true.}

\item{remove_non_independent_records}{logical. option to control including or removing records that have been assessed to be non independent. Default is set to true.}

\item{min_delta_time}{integer. integer value specifying independence of records, if set to 0, all images are independent and no temporal filtering is needed. Default is set to 3.}
}
\value{
Generates a dataframe with species temporal information for the all camera traps and stations that are present in the trip.
}
\description{
Generate a data frame that is compatible with the camtrapR recordTable data structure.
}
