% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recocam_various_helper_functions.R
\name{make_recordtable_helper}
\alias{make_recordtable_helper}
\title{Construct a recordTable, data structure that is compatible with camtrapR for only species image.}
\usage{
make_recordtable_helper(
  df_with_species_images,
  delta_time_compared_to,
  camera_independence,
  remove_non_indpendent_records,
  min_delta_time
)
}
\arguments{
\item{df_with_species_images}{R dataframe ofonly images that have species in them, with key attributes - trip, site, camera_no, site_code, moon_phase_number, species, count.}

\item{delta_time_compared_to}{Character string having values of lastIndependentRecord (default) or lastRecord.
This is used to compare the time difference to previous records of same species in a station.}

\item{camera_independence}{Logical value with default set to TRUE.}

\item{remove_non_indpendent_records}{Logical value with default set to TRUE.}

\item{min_delta_time}{Integer value specifying independence of records, if set to 0,
all images are independent and no temporal filtering is needed. Default is set to 3.}
}
\value{
R Data frame which has all the attributes of cameratrapR recordTable data structure.
}
\description{
Construct a recordTable, data structure that is compatible with camtrapR for only species image.
}
